+++ src/@next/pages/CheckoutPage/subpages/CheckoutPaymentSubpage.tsx
@@ -1,31 +1,26 @@
 import React, {
   forwardRef,
   RefForwardingComponent,
-  useContext,
-  useEffect,
   useImperativeHandle,
   useRef,
   useState,
 } from "react";
 import { useIntl } from "react-intl";
-import { RouteComponentProps, useHistory } from "react-router";
+import { RouteComponentProps } from "react-router";
 
 import { CheckoutPayment } from "@components/organisms";
-import { useAuth, useCart, useCheckout } from "@saleor/sdk";
-import { ShopContext } from "@temp/components/ShopProvider/context";
-import { CHECKOUT_STEPS } from "@temp/core/config";
+import { useCheckout } from "@saleor/sdk";
 import { commonMessages } from "@temp/intl";
-import { IAddress, ICardData, IFormError } from "@types";
-import { filterNotEmptyArrayItems } from "@utils/misc";
+import { IFormError } from "@types";
 
 export interface ICheckoutPaymentSubpageHandles {
   submitPayment: () => void;
 }
 interface IProps extends RouteComponentProps<any> {
-  selectedPaymentGateway?: string;
-  selectedPaymentGatewayToken?: string;
-  selectPaymentGateway: (paymentGateway: string) => void;
+  paymentGatewayFormRef: React.RefObject<HTMLFormElement>;
   changeSubmitProgress: (submitInProgress: boolean) => void;
+  onSubmitSuccess: () => void;
+  onPaymentGatewayError: (errors: IFormError[]) => void;
 }
 
 const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
@@ -33,168 +28,41 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
   IProps
 > = (
   {
-    selectedPaymentGateway,
-    selectedPaymentGatewayToken,
+    paymentGatewayFormRef,
     changeSubmitProgress,
-    selectPaymentGateway,
+    onSubmitSuccess,
+    onPaymentGatewayError,
     ...props
   }: IProps,
   ref
 ) => {
-  const history = useHistory();
-  const { user } = useAuth();
-  const {
-    checkout,
-    billingAsShipping,
-    setBillingAddress,
-    setBillingAsShippingAddress,
-    selectedBillingAddressId,
-    availablePaymentGateways,
-    promoCodeDiscount,
-    addPromoCode,
-    removePromoCode,
-    createPayment,
-  } = useCheckout();
-  const { items } = useCart();
-  const { countries } = useContext(ShopContext);
+  const { promoCodeDiscount, addPromoCode, removePromoCode } = useCheckout();
 
-  const isShippingRequiredForProducts =
-    items &&
-    items.some(
-      ({ variant }) => variant.product?.productType.isShippingRequired
-    );
-
-  const [billingErrors, setBillingErrors] = useState<IFormError[]>([]);
-  const [gatewayErrors, setGatewayErrors] = useState<IFormError[]>([]);
   const [promoCodeErrors, setPromoCodeErrors] = useState<IFormError[]>([]);
 
-  const [billingAsShippingState, setBillingAsShippingState] = useState(
-    billingAsShipping
-  );
-  useEffect(() => {
-    setBillingAsShippingState(billingAsShipping);
-  }, [billingAsShipping]);
-
-  const checkoutBillingAddress = checkout?.billingAddress
-    ? {
-        ...checkout?.billingAddress,
-        phone: checkout?.billingAddress?.phone || undefined,
-      }
-    : undefined;
-  const paymentGateways = availablePaymentGateways || [];
-
-  const checkoutBillingFormId = "billing-form";
-  const checkoutBillingFormRef = useRef<HTMLFormElement>(null);
-  const checkoutGatewayFormId = "gateway-form";
-  const checkoutGatewayFormRef = useRef<HTMLFormElement>(null);
-  const checkoutNewAddressFormId = "new-address-form";
   const promoCodeDiscountFormId = "discount-form";
   const promoCodeDiscountFormRef = useRef<HTMLFormElement>(null);
   const intl = useIntl();
 
   useImperativeHandle(ref, () => ({
     submitPayment: () => {
-      if (billingAsShippingState) {
-        handleSetBillingAddress();
-      } else if (user && selectedBillingAddressId) {
-        checkoutBillingFormRef.current?.dispatchEvent(
-          new Event("submit", { cancelable: true })
-        );
-      } else {
-        // TODO validate form
-        checkoutBillingFormRef.current?.dispatchEvent(
-          new Event("submit", { cancelable: true })
-        );
-      }
-    },
-  }));
-
-  const handleProcessPayment = async (
-    gateway: string,
-    token: string,
-    cardData?: ICardData
-  ) => {
-    const { dataError } = await createPayment(gateway, token, cardData);
-    const errors = dataError?.error;
-    changeSubmitProgress(false);
-    if (errors) {
-      setGatewayErrors(errors);
-    } else {
-      setGatewayErrors([]);
-      history.push(CHECKOUT_STEPS[2].nextStepLink);
-    }
-  };
-  const handlePaymentGatewayError = () => {
-    changeSubmitProgress(false);
-  };
-  const handleSetBillingAddress = async (
-    address?: IAddress,
-    email?: string,
-    userAddressId?: string
-  ) => {
-    if (!address && !billingAsShippingState) {
-      setBillingErrors([
-        {
-          message: intl.formatMessage({
-            defaultMessage: "Please provide billing address.",
-          }),
-        },
-      ]);
-      return;
-    }
-
-    const billingEmail = user?.email || email;
-
-    if (
-      !billingEmail &&
-      !billingAsShippingState &&
-      !isShippingRequiredForProducts
-    ) {
-      setBillingErrors([
-        {
-          field: "email",
-          message: intl.formatMessage(commonMessages.provideEmailAddress),
-        },
-      ]);
-      return;
-    }
-
-    let errors;
-    changeSubmitProgress(true);
-    if (billingAsShippingState && isShippingRequiredForProducts) {
-      const { dataError } = await setBillingAsShippingAddress();
-      errors = dataError?.error;
-    } else {
-      const { dataError } = await setBillingAddress(
-        {
-          ...address,
-          id: userAddressId,
-        },
-        billingEmail
-      );
-      errors = dataError?.error;
-    }
-    if (errors) {
-      changeSubmitProgress(false);
-      setBillingErrors(errors);
-    } else {
-      setBillingErrors([]);
       if (promoCodeDiscountFormRef.current) {
         promoCodeDiscountFormRef.current?.dispatchEvent(
           new Event("submit", { cancelable: true })
         );
-      } else if (checkoutGatewayFormRef.current) {
-        checkoutGatewayFormRef.current.dispatchEvent(
+      } else if (paymentGatewayFormRef.current) {
+        paymentGatewayFormRef.current.dispatchEvent(
           new Event("submit", { cancelable: true })
         );
       } else {
         changeSubmitProgress(false);
-        setGatewayErrors([
+        onPaymentGatewayError([
           { message: intl.formatMessage(commonMessages.choosePaymentMethod) },
         ]);
       }
-    }
-  };
+    },
+  }));
+
   const handleAddPromoCode = async (promoCode: string) => {
     const { dataError } = await addPromoCode(promoCode);
     const errors = dataError?.error;
@@ -203,13 +71,13 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
       setPromoCodeErrors(errors);
     } else {
       setPromoCodeErrors([]);
-      if (checkoutGatewayFormRef.current) {
-        checkoutGatewayFormRef.current.dispatchEvent(
+      if (paymentGatewayFormRef.current) {
+        paymentGatewayFormRef.current.dispatchEvent(
           new Event("submit", { cancelable: true })
         );
       } else {
         changeSubmitProgress(false);
-        setGatewayErrors([
+        onPaymentGatewayError([
           { message: intl.formatMessage(commonMessages.choosePaymentMethod) },
         ]);
       }
@@ -223,26 +91,26 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
       setPromoCodeErrors(errors);
     } else {
       setPromoCodeErrors([]);
-      if (checkoutGatewayFormRef.current) {
-        checkoutGatewayFormRef.current.dispatchEvent(
+      if (paymentGatewayFormRef.current) {
+        paymentGatewayFormRef.current.dispatchEvent(
           new Event("submit", { cancelable: true })
         );
       } else {
         changeSubmitProgress(false);
-        setGatewayErrors([
+        onPaymentGatewayError([
           { message: intl.formatMessage(commonMessages.choosePaymentMethod) },
         ]);
       }
     }
   };
   const handleSubmitUnchangedDiscount = () => {
-    if (checkoutGatewayFormRef.current) {
-      checkoutGatewayFormRef.current.dispatchEvent(
+    if (paymentGatewayFormRef.current) {
+      paymentGatewayFormRef.current.dispatchEvent(
         new Event("submit", { cancelable: true })
       );
     } else {
       changeSubmitProgress(false);
-      setGatewayErrors([
+      onPaymentGatewayError([
         { message: intl.formatMessage(commonMessages.choosePaymentMethod) },
       ]);
     }
@@ -251,36 +119,6 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
   return (
     <CheckoutPayment
       {...props}
-      billingErrors={billingErrors}
-      gatewayErrors={gatewayErrors}
-      billingFormId={checkoutBillingFormId}
-      billingFormRef={checkoutBillingFormRef}
-      userAddresses={user?.addresses
-        ?.filter(filterNotEmptyArrayItems)
-        .map(
-          ({
-            isDefaultBillingAddress,
-            isDefaultShippingAddress,
-            phone,
-            ...address
-          }) => ({
-            ...address,
-            isDefaultBillingAddress: !!isDefaultBillingAddress,
-            isDefaultShippingAddress: !!isDefaultShippingAddress,
-            phone: phone || undefined,
-          })
-        )}
-      selectedUserAddressId={selectedBillingAddressId}
-      checkoutBillingAddress={checkoutBillingAddress}
-      countries={countries}
-      paymentGateways={paymentGateways}
-      selectedPaymentGateway={selectedPaymentGateway}
-      selectedPaymentGatewayToken={selectedPaymentGatewayToken}
-      selectPaymentGateway={selectPaymentGateway}
-      setBillingAddress={handleSetBillingAddress}
-      billingAsShippingPossible={!!isShippingRequiredForProducts}
-      billingAsShippingAddress={billingAsShippingState}
-      setBillingAsShippingAddress={setBillingAsShippingState}
       promoCodeDiscountFormId={promoCodeDiscountFormId}
       promoCodeDiscountFormRef={promoCodeDiscountFormRef}
       promoCodeDiscount={{
@@ -290,12 +128,6 @@ const CheckoutPaymentSubpageWithRef: RefForwardingComponent<
       removeVoucherCode={handleRemovePromoCode}
       submitUnchangedDiscount={handleSubmitUnchangedDiscount}
       promoCodeErrors={promoCodeErrors}
-      gatewayFormId={checkoutGatewayFormId}
-      gatewayFormRef={checkoutGatewayFormRef}
-      userId={user?.id}
-      newAddressFormId={checkoutNewAddressFormId}
-      processPayment={handleProcessPayment}
-      onGatewayError={handlePaymentGatewayError}
     />
   );
 };